+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/table/FilterTable.java
@@ -915,7 +915,7 @@ public final class FilterTable<R, C> extends JTable {
 										builder.cellRendererFactory.create(column.identifier(), this))));
 		columnModel().columns().stream()
 						.filter(column -> column.getHeaderRenderer() == null)
-						.forEach(column -> column.setHeaderRenderer(builder.headerRendererFactory.create(column, tableModel)));
+						.forEach(column -> column.setHeaderRenderer(builder.headerRendererFactory.create(column, this)));
 		columnModel().columns().stream()
 						.filter(column -> column.getCellEditor() == null)
 						.forEach(column -> {
@@ -1277,7 +1277,7 @@ public final class FilterTable<R, C> extends JTable {
 		 * @param headerRendererFactory the header renderer factory
 		 * @return this builder instance
 		 */
-		Builder<R, C> headerRendererFactory(FilterTableHeaderRenderer.Factory<C> headerRendererFactory);
+		Builder<R, C> headerRendererFactory(FilterTableHeaderRenderer.Factory<R, C> headerRendererFactory);
 
 		/**
 		 * the cell renderer for the given column, overrides {@link #cellEditorFactory(FilterTableCellEditor.Factory)}.
@@ -1505,7 +1505,7 @@ public final class FilterTable<R, C> extends JTable {
 		private SummaryValues.@Nullable Factory<C> summaryValuesFactory;
 		private TableConditionPanel.Factory<C> filterPanelFactory = new DefaultFilterPanelFactory<>();
 		private ComponentFactory filterComponentFactory = new FilterComponentFactory();
-		private FilterTableHeaderRenderer.Factory<C> headerRendererFactory;
+		private FilterTableHeaderRenderer.Factory<R, C> headerRendererFactory;
 		private FilterTableCellRenderer.Factory<R, C> cellRendererFactory;
 		private FilterTableCellEditor.@Nullable Factory<C> cellEditorFactory;
 		private BiPredicate<R, C> cellEditable = (BiPredicate<R, C>) CELL_EDITABLE;
@@ -1590,7 +1590,7 @@ public final class FilterTable<R, C> extends JTable {
 		}
 
 		@Override
-		public Builder<R, C> headerRendererFactory(FilterTableHeaderRenderer.Factory<C> headerRendererFactory) {
+		public Builder<R, C> headerRendererFactory(FilterTableHeaderRenderer.Factory<R, C> headerRendererFactory) {
 			this.headerRendererFactory = requireNonNull(headerRendererFactory);
 			return this;
 		}